# worker/contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Planner & Task Scheduler API
  version: 1.1.0
  description: Authoritative OpenAPI 3.1 contract for Projects, Tasks, and Scheduler Automation.
paths:
  /api/v1/projects:
    post:
      summary: Create or upsert a project
      operationId: upsertProject
      parameters:
        - name: x-request-id
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertProjectRequest'
      responses:
        '200':
          description: Project successfully upserted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSingleResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List projects (paginated)
      operationId: listProjects
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [ACTIVE, ON_HOLD, COMPLETED, ARCHIVED]
      responses:
        '200':
          description: Paginated list of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'

  /api/v1/projects/{projectId}:
    get:
      summary: Get project details by ID
      operationId: getProjectById
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSingleResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update project metadata or status
      operationId: updateProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSingleResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a project
      operationId: deleteProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/projects/{projectId}/tasks:
    post:
      summary: Upsert a task within a project
      operationId: upsertTask
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: x-request-id
          in: header
          required: true
          schema:
            type: string
        - name: x-idempotency-key
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertTaskRequest'
      responses:
        '200':
          description: Task successfully upserted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSingleResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List tasks for a project (paginated)
      operationId: listTasks
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, IN_PROGRESS, COMPLETED, CANCELLED]
      responses:
        '200':
          description: Paginated list of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'

  /api/v1/projects/{projectId}/tasks/{taskId}:
    get:
      summary: Get a specific task by ID
      operationId: getTaskById
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSingleResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a specific task
      operationId: deleteTask
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/projects/{projectId}/tasks/{taskId}/status:
    patch:
      summary: Update task status
      operationId: updateTaskStatus
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskStatusRequest'
      responses:
        '200':
          description: Task status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSingleResponse'

  /api/v1/scheduler/trigger:
    post:
      summary: Trigger immediate sweep for pending tasks
      operationId: triggerSchedulerSweep
      responses:
        '200':
          description: Scheduler sweep completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerSweepResponse'

components:
  schemas:
    PaginationMeta:
      type: object
      required:
        - page
        - pageSize
        - totalItems
        - totalPages
        - hasNextPage
        - hasPreviousPage
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
        hasNextPage:
          type: boolean
        hasPreviousPage:
          type: boolean
        nextCursor:
          type: string
          nullable: true

    Meta:
      type: object
      required:
        - requestId
        - correlationId
        - timestamp
        - executionTimeMs
        - apiVersion
      properties:
        requestId:
          type: string
        correlationId:
          type: string
        timestamp:
          type: string
        executionTimeMs:
          type: integer
        apiVersion:
          type: string
        pagination:
          $ref: '#/components/schemas/PaginationMeta'

    Project:
      type: object
      required:
        - projectId
        - name
        - status
        - createdAt
        - updatedAt
      properties:
        projectId:
          type: string
        name:
          type: string
        description:
          type: string
        ownerEmail:
          type: string
        status:
          type: string
          enum: [ACTIVE, ON_HOLD, COMPLETED, ARCHIVED]
        createdAt:
          type: string
        updatedAt:
          type: string

    UpsertProjectRequest:
      type: object
      required:
        - projectId
        - name
      properties:
        projectId:
          type: string
        name:
          type: string
        description:
          type: string
        ownerEmail:
          type: string
        status:
          type: string
          enum: [ACTIVE, ON_HOLD, COMPLETED, ARCHIVED]

    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        ownerEmail:
          type: string
        status:
          type: string
          enum: [ACTIVE, ON_HOLD, COMPLETED, ARCHIVED]

    ProjectSingleResponse:
      type: object
      required:
        - success
        - statusCode
        - data
        - meta
      properties:
        success:
          type: boolean
        statusCode:
          type: integer
        data:
          $ref: '#/components/schemas/Project'
        meta:
          $ref: '#/components/schemas/Meta'

    ProjectListResponse:
      type: object
      required:
        - success
        - statusCode
        - data
        - meta
      properties:
        success:
          type: boolean
        statusCode:
          type: integer
        data:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        meta:
          $ref: '#/components/schemas/Meta'

    Task:
      type: object
      required:
        - taskId
        - projectId
        - title
        - status
        - createdAt
        - updatedAt
      properties:
        taskId:
          type: string
        projectId:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, COMPLETED, CANCELLED]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        assignedToEmail:
          type: string
        dueDate:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string

    UpsertTaskRequest:
      type: object
      required:
        - taskId
        - title
      properties:
        taskId:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, COMPLETED, CANCELLED]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        assignedToEmail:
          type: string
        dueDate:
          type: string

    UpdateTaskStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, COMPLETED, CANCELLED]

    TaskSingleResponse:
      type: object
      required:
        - success
        - statusCode
        - data
        - meta
      properties:
        success:
          type: boolean
        statusCode:
          type: integer
        data:
          $ref: '#/components/schemas/Task'
        meta:
          $ref: '#/components/schemas/Meta'

    TaskListResponse:
      type: object
      required:
        - success
        - statusCode
        - data
        - meta
      properties:
        success:
          type: boolean
        statusCode:
          type: integer
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        meta:
          $ref: '#/components/schemas/Meta'

    DeleteResponse:
      type: object
      required:
        - success
        - statusCode
        - data
        - meta
      properties:
        success:
          type: boolean
        statusCode:
          type: integer
        data:
          type: object
        meta:
          $ref: '#/components/schemas/Meta'

    ErrorResponse:
      type: object
      required:
        - success
        - statusCode
        - error
        - meta
      properties:
        success:
          type: boolean
          enum: [false]
        statusCode:
          type: integer
        error:
          type: object
        meta:
          $ref: '#/components/schemas/Meta'

    SchedulerSweepResponse:
      type: object
      required:
        - success
        - statusCode
        - data
        - meta
      properties:
        success:
          type: boolean
        statusCode:
          type: integer
        data:
          type: object
        meta:
          $ref: '#/components/schemas/Meta'